Fix Pget_bkpt_ip() return value.
This function returns an arbitrary pointer, so in order to be able to
represent addresses in the upper half of the address space it must
return a uintptr_t, not a long.  Two unlikely addresses remain
unrepresentable: zero (no breakpoint) and (uintptr_t)-1 (error).

All arch implementations already conform to this contract: all we need
to do is fix a few declarations.

+++ libproc/Pcontrol.h
@@ -284,7 +284,7 @@ extern	void	Psym_free(struct ps_prochandle *);
 extern	void	Psym_release(struct ps_prochandle *);
 extern	int	Pread_isa_info(struct ps_prochandle *P, const char *procname);
 extern	void	Preadauxvec(struct ps_prochandle *P);
-extern  long	Pget_bkpt_ip(struct ps_prochandle *P, int expect_esrch);
+extern  uintptr_t Pget_bkpt_ip(struct ps_prochandle *P, int expect_esrch);
 extern  long	Preset_bkpt_ip(struct ps_prochandle *P, uintptr_t addr);
 extern	char *	Pget_proc_status(pid_t pid, const char *field);
