+++ libproc/sparc/isadep.h
@@ -23,7 +23,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2013 Oracle, Inc.  All rights reserved.
+ * Copyright 2013, 2015 Oracle, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -51,7 +51,7 @@ isa_dispatch_t dispatch[] = {
 
 #ifdef WANT_GET_BKPT_IP
 
-extern	long Pget_bkpt_ip_sparc64(struct ps_prochandle *P, int expect_esrch);
+extern uintptr_t Pget_bkpt_ip_sparc64(struct ps_prochandle *P, int expect_esrch);
 
 /*
  * I hope 32-on-64 traps are treated like 64-on-64 traps.
